refactor(serve): the served registry's tests in their own file
The file cap (1500 lines) caught the serve tests at 1531 after the
scope test of wave 7.d: it lives in its own module now, next to the
schedule tests, with the fixtures it shares made visible to siblings.
Nothing changes in what the tests say.
